Move common test utils
