+++ src/common/snippets/tests/CMakeLists.txt
@@ -12,7 +12,7 @@ addIeTargetTest(
             $<TARGET_PROPERTY:inference_engine_snippets,INTERFACE_INCLUDE_DIRECTORIES>
         LINK_LIBRARIES
             openvino::runtime::dev
-            commonTestUtils
+            common_test_utils
             snippetsNgraphFunctions
         ADD_CPPLINT
         LABELS
@@ -30,4 +30,4 @@ ie_faster_build(${TARGET_NAME}
 add_library(snippets_test_utils STATIC ${CMAKE_CURRENT_SOURCE_DIR}/include/lowering_utils.hpp
                                        ${CMAKE_CURRENT_SOURCE_DIR}/src/lowering_utils.cpp)
 target_include_directories(snippets_test_utils PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
-target_link_libraries(snippets_test_utils PRIVATE commonTestUtils snippetsNgraphFunctions)
+target_link_libraries(snippets_test_utils PRIVATE common_test_utils snippetsNgraphFunctions)
