+++ src/core/tests/CMakeLists.txt
@@ -39,7 +39,7 @@ ov_add_test_target(
             test_model_zoo
         LINK_LIBRARIES
             engines_test_util
-            commonTestUtils
+            common_test_utils
             ngraph_reference
             ngraph::builder
             openvino::util
