+++ src/frontends/ir/tests/CMakeLists.txt
@@ -13,7 +13,7 @@ ov_add_test_target(
             gtest
             gtest_main
             openvino::runtime::dev
-            commonTestUtils
+            common_test_utils
         INCLUDES
             "${CMAKE_CURRENT_SOURCE_DIR}/../include"
         ADD_CLANG_FORMAT
