+++ src/frontends/tests/frontend/shared/CMakeLists.txt
@@ -18,7 +18,7 @@ target_include_directories(${TARGET_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../
 target_link_libraries(${TARGET_NAME}
         PUBLIC
                 engines_test_util offline_transformations
-                commonTestUtils ngraph_test_util funcTestUtils openvino::util
+                common_test_utils ngraph_test_util funcTestUtils openvino::util
                 openvino::runtime
         PRIVATE
                 cnpy)
