+++ src/inference/tests/functional/caching_test.cpp
@@ -28,7 +28,6 @@
 #include "openvino/op/logical_not.hpp"
 #include "openvino/util/file_util.hpp"
 #include "unit_test_utils/mocks/cpp_interfaces/interface/mock_iexecutable_network_internal.hpp"
-#include "unit_test_utils/mocks/cpp_interfaces/interface/mock_iinference_plugin.hpp"
 #include "unit_test_utils/mocks/mock_iexecutable_network.hpp"
 #include "unit_test_utils/mocks/mock_iinfer_request.hpp"
