+++ src/plugins/auto_batch/tests/unit/CMakeLists.txt
@@ -4,6 +4,10 @@
 
 set(TARGET_NAME ov_auto_batch_unit_tests)
 
+if(SUGGEST_OVERRIDE_SUPPORTED)
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-suggest-override")
+endif()
+
 set(CI_BUILD_NUMBER "unittest")
 add_definitions(-DAUTOBATCH_UNITTEST)
