+++ src/plugins/intel_gna/tests/deprecated/helpers/CMakeLists.txt
@@ -30,12 +30,12 @@ endfunction()
 
 add_helpers(${TARGET_NAME})
 
-target_link_libraries(${TARGET_NAME} PUBLIC commonTestUtils
+target_link_libraries(${TARGET_NAME} PUBLIC common_test_utils
                                      PRIVATE inference_engine_legacy)
 
 add_helpers(${TARGET_NAME}_s USE_STATIC_IE)
 
-target_link_libraries(${TARGET_NAME}_s PUBLIC commonTestUtils_s
+target_link_libraries(${TARGET_NAME}_s PUBLIC common_test_utils_s
                                        PRIVATE inference_engine_legacy_s)
 
 if (ENABLE_DATA)
