+++ src/tests/CMakeLists.txt
@@ -28,8 +28,14 @@ ov_deprecated_no_errors()
 
 add_subdirectory(ngraph_helpers)
 add_subdirectory(ie_test_utils)
+add_subdirectory(test_utils)
 add_subdirectory(util)
 add_subdirectory(engines_util)
+
+
+if(ENABLE_TESTS)
+    add_subdirectory(util_tests)
+endif()
 
 if(ENABLE_FUNCTIONAL_TESTS)
     add_subdirectory(functional)
