+++ src/tests/functional/plugin/conformance/subgraphs_dumper/include/matchers/base_matcher.hpp
@@ -9,7 +9,7 @@
 #include "ngraph/node.hpp"
 #include "pugixml.hpp"
 
-#include "functional_test_utils/include/functional_test_utils/summary/op_info.hpp"
+#include "functional_test_utils/summary/op_info.hpp"
 
 namespace SubgraphsDumper {
