+++ src/tests/functional/plugin/conformance/subgraphs_dumper/tests/matchers/convolutions_matcher.cpp
@@ -5,7 +5,7 @@
 #include "gtest/gtest.h"
 #include "matchers/convolutions.hpp"
 #include "ngraph/ops.hpp"
-#include "functional_test_utils/include/functional_test_utils/summary/op_info.hpp"
+#include "functional_test_utils/summary/op_info.hpp"
 
 class ConvolutionMatcherTest : public ::testing::Test {
 protected:
