+++ src/tests/ie_test_utils/unit_test_utils/mocks/cpp_interfaces/impl/mock_inference_plugin_internal.hpp
@@ -13,23 +13,13 @@
 #include <vector>
 #include <cpp_interfaces/interface/ie_iplugin_internal.hpp>
 #include <cpp_interfaces/interface/ie_iexecutable_network_internal.hpp>
-
-class MockInferencePluginInternal2 : public InferenceEngine::IInferencePlugin {
-public:
-    MOCK_METHOD2(LoadExeNetworkImpl, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
-            const InferenceEngine::CNNNetwork &, const std::map<std::string, std::string> &));
-    MOCK_METHOD2(LoadNetwork, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
-            const InferenceEngine::CNNNetwork &,
-            const std::map<std::string, std::string> &));
-    MOCK_METHOD1(AddExtension, void(InferenceEngine::IExtensionPtr ext_ptr));
-    MOCK_METHOD1(SetConfig, void(const std::map <std::string, std::string> &));
-};
+#include "openvino/runtime/iplugin.hpp"
 
 class MockInferencePluginInternal : public InferenceEngine::IInferencePlugin {
 public:
     MOCK_METHOD2(LoadExeNetworkImpl, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
             const InferenceEngine::CNNNetwork &, const std::map<std::string, std::string> &));
-    MOCK_METHOD1(AddExtension, void(InferenceEngine::IExtensionPtr ext_ptr));
+    MOCK_METHOD1(AddExtension, void(const std::shared_ptr<InferenceEngine::IExtension>&));
     MOCK_METHOD1(SetConfig, void(const std::map <std::string, std::string> &));
 
     std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>
@@ -38,12 +28,4 @@ public:
     }
 
     std::string importedString;
-};
-
-class MockInferencePluginInternal3 : public InferenceEngine::IInferencePlugin {
-public:
-    MOCK_METHOD2(LoadExeNetworkImpl, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
-            const InferenceEngine::CNNNetwork &, const std::map<std::string, std::string> &));
-    MOCK_METHOD1(AddExtension, void(InferenceEngine::IExtensionPtr ext_ptr));
-    MOCK_METHOD1(SetConfig, void(const std::map <std::string, std::string> &));
 };
