+++ src/tests/ie_test_utils/unit_test_utils/mocks/cpp_interfaces/interface/mock_iinference_plugin.hpp
@@ -13,7 +13,7 @@
 
 class MockIInferencePlugin : public InferenceEngine::IInferencePlugin {
 public:
-    MOCK_METHOD1(AddExtension, void(InferenceEngine::IExtensionPtr));
+    MOCK_METHOD1(AddExtension, void(const std::shared_ptr<InferenceEngine::IExtension>&));
     MOCK_METHOD2(LoadNetwork, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
                 const InferenceEngine::CNNNetwork&, const std::map<std::string, std::string>&));
     MOCK_METHOD2(LoadNetwork, ov::SoPtr<InferenceEngine::IExecutableNetworkInternal>(
@@ -36,7 +36,7 @@ public:
     MOCK_METHOD1(GetDefaultContext, std::shared_ptr<InferenceEngine::RemoteContext>(const InferenceEngine::ParamMap&));
     MOCK_METHOD3(LoadNetwork, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
                 const InferenceEngine::CNNNetwork&, const std::map<std::string, std::string>&,
-                std::shared_ptr<InferenceEngine::RemoteContext>));
+                const std::shared_ptr<InferenceEngine::RemoteContext>&));
     MOCK_METHOD2(ImportNetwork, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
                 std::istream&, const std::map<std::string, std::string>&));
     MOCK_METHOD3(ImportNetwork, std::shared_ptr<InferenceEngine::IExecutableNetworkInternal>(
