+++ src/tests/ngraph_helpers/snippets_ngraph_functions/CMakeLists.txt
@@ -6,7 +6,7 @@ set(TARGET_NAME snippetsNgraphFunctions)
 
 set(PUBLIC_HEADERS_DIR "${CMAKE_CURRENT_SOURCE_DIR}/include")
 set(SNIPPETS_INCLUDES "$<TARGET_PROPERTY:inference_engine_snippets,SOURCE_DIR>/include")
-set(COMMON_TEST_UTILS_INCLUDES "$<TARGET_PROPERTY:commonTestUtils,INTERFACE_INCLUDE_DIRECTORIES>")
+set(COMMON_TEST_UTILS_INCLUDES "$<TARGET_PROPERTY:common_test_utils,INTERFACE_INCLUDE_DIRECTORIES>")
 addIeTarget(
         NAME ${TARGET_NAME}
         TYPE STATIC
@@ -23,7 +23,7 @@ addIeTarget(
         LINK_LIBRARIES
             PUBLIC
                 openvino::runtime::dev
-                commonTestUtils
+                common_test_utils
                 inference_engine_snippets
                 lptNgraphFunctions
         ADD_CPPLINT
