+++ src/tests/test_utils/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Copyright (C) 2018-2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+if(ENABLE_TESTS)
+    add_subdirectory(common_test_utils)
+endif()
