+++ src/tests/test_utils/common_test_utils/CMakeLists.txt
@@ -8,9 +8,6 @@ function(add_common_utils ADD_TARGET_NAME)
             NAME ${ADD_TARGET_NAME}
             TYPE STATIC
             ROOT ${CMAKE_CURRENT_SOURCE_DIR}
-            EXCLUDED_SOURCE_PATHS
-                ${CMAKE_CURRENT_SOURCE_DIR}/gtest
-                ${CMAKE_CURRENT_SOURCE_DIR}/tests
             ADD_CPPLINT
             DEVELOPER_PACKAGE
                 tests
@@ -24,6 +21,10 @@ function(add_common_utils ADD_TARGET_NAME)
                     openvino::runtime::dev
                 PRIVATE
                     openvino::util
+                INCLUDES
+                    PUBLIC
+                        "${CMAKE_CURRENT_SOURCE_DIR}/include"
+
     )
 
     # USE_STATIC_IE is passed
@@ -33,7 +34,7 @@ function(add_common_utils ADD_TARGET_NAME)
 
     ie_faster_build(${ADD_TARGET_NAME}
         UNITY
-        PCH PRIVATE "precomp.hpp"
+        PCH PRIVATE "include/common_test_utils/precomp.hpp"
     )
 
     # detecting regex support
@@ -62,9 +63,13 @@ function(add_common_utils ADD_TARGET_NAME)
     target_compile_definitions(${ADD_TARGET_NAME} PUBLIC ${ARGN})
 endfunction()
 
+# Keep old name so that library can be used from VPU repo
 set(TARGET_NAME commonTestUtils)
+set(NEW_TARGET_NAME common_test_utils)
 
 add_common_utils(${TARGET_NAME})
 add_common_utils(${TARGET_NAME}_s USE_STATIC_IE)
 
-add_subdirectory(tests)
+# Add new names to use them from openvino repo
+add_library(${NEW_TARGET_NAME} ALIAS ${TARGET_NAME})
+add_library(${NEW_TARGET_NAME}_s ALIAS ${TARGET_NAME}_s)
