+++ src/tests/test_utils/common_test_utils/include/common_test_utils/ngraph_test_utils.hpp
@@ -8,12 +8,12 @@
 #include <memory>
 #include <queue>
 
-#include "graph_comparator.hpp"
+#include "common_test_utils/graph_comparator.hpp"
 #include "openvino/core/dimension.hpp"
 #include "openvino/core/model.hpp"
 #include "openvino/pass/manager.hpp"
 #include "openvino/pass/pass.hpp"
-#include "test_common.hpp"
+#include "common_test_utils/test_common.hpp"
 #include "transformations/init_node_info.hpp"
 
 #define DYN ov::Dimension::dynamic()
