+++ src/tests/test_utils/common_test_utils/include/common_test_utils/test_common.hpp
@@ -7,7 +7,7 @@
 #include <gtest/gtest.h>
 
 #include <string>
-#include "test_assertions.hpp"
+#include "common_test_utils/test_assertions.hpp"
 
 namespace CommonTestUtils {
