+++ src/tests/test_utils/common_test_utils/include/common_test_utils/unicode_utils.hpp
@@ -10,8 +10,8 @@
 #include <algorithm>
 
 #include <file_utils.h>
-#include "common_utils.hpp"
-#include "w_dirent.h"
+#include "common_test_utils/common_utils.hpp"
+#include "common_test_utils/w_dirent.h"
 
 #ifdef OPENVINO_ENABLE_UNICODE_PATH_SUPPORT
