+++ src/tests/test_utils/common_test_utils/src/ngraph_test_utils.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "ngraph_test_utils.hpp"
+#include "common_test_utils/ngraph_test_utils.hpp"
 
 namespace ov {
 namespace pass {
