+++ src/tests/test_utils/common_test_utils/src/test_common.cpp
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "test_common.hpp"
-#include "common_utils.hpp"
-#include "test_constants.hpp"
+#include "common_test_utils/test_common.hpp"
+#include "common_test_utils/common_utils.hpp"
+#include "common_test_utils/test_constants.hpp"
 
 #include <threading/ie_executor_manager.hpp>
