+++ src/tests/util_tests/CMakeLists.txt
@@ -2,14 +2,13 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-set(TARGET_NAME commonUtilsTests)
+set(TARGET_NAME ov_util_tests)
 
 addIeTargetTest(
         NAME ${TARGET_NAME}
         ROOT ${CMAKE_CURRENT_SOURCE_DIR}
         DEPENDENCIES
-        commonTestUtils
         LINK_LIBRARIES
-        commonTestUtils
+          common_test_utils
         ADD_CPPLINT
 )
